refactor: rename embeddings.rs to type_aliases.rs
