+++ swiftide/src/lib.rs
@@ -1,9 +1,9 @@
-pub mod embeddings;
 pub mod ingestion;
 pub mod integrations;
 pub mod loaders;
 pub mod traits;
 pub mod transformers;
+pub mod type_aliases;
 
-pub use embeddings::*;
 pub use traits::*;
+pub use type_aliases::*;
