+++ swiftide/src/transformers/mod.rs
@@ -1,10 +1,14 @@
+#[cfg(feature = "tree-sitter")]
 pub mod chunk_code;
+
 pub mod chunk_markdown;
 pub mod embed;
 pub mod metadata_qa_code;
 pub mod metadata_qa_text;
 
+#[cfg(feature = "tree-sitter")]
 pub use chunk_code::ChunkCode;
+
 pub use chunk_markdown::ChunkMarkdown;
 pub use embed::Embed;
 pub use metadata_qa_code::MetadataQACode;
